Make get_exec_output() work on big endian systems.
Undo API breakage introduced by commit 7374b84fe4c2ef8497fde3dae80a69aa89eba19c.

+++ agent/mibgroup/util_funcs.c
@@ -244,7 +244,7 @@ get_exec_output(struct extensible *ex)
 #if HAVE_EXECV
     char            cachefile[STRMAX];
     char            cache[NETSNMP_MAXCACHESIZE];
-    size_t          cachebytes;
+    int             cachebytes;
     int             cfd;
 #ifdef NETSNMP_EXCACHETIME
     long            curtime;
